moved views related to Group elements pages to separate file
